chore:  move  logic into handlers via dispatcher
